refactor: Payment DTO 구조 개선 (Request/Response 분리)
