+++ backend/demo/src/main/java/com/sesac/solbid/dto/payment/request/PaymentApproveRequest.java
@@ -1,4 +1,4 @@
-package com.sesac.solbid.dto.request;
+package com.sesac.solbid.dto.payment.request;
 
 import lombok.Getter;
 import lombok.Setter;
