+++ backend/demo/src/main/java/com/sesac/solbid/dto/payment/response/PaymentPrepareResponse.java
@@ -1,4 +1,4 @@
-package com.sesac.solbid.dto.response;
+package com.sesac.solbid.dto.payment.response;
 
 
 import lombok.Data;
