+++ backend/demo/src/main/java/com/sesac/solbid/service/PaymentService.java
@@ -1,8 +1,7 @@
 package com.sesac.solbid.service;
 
-import com.sesac.solbid.dto.request.PaymentPrepareRequest;
-import com.sesac.solbid.dto.response.PaymentPrepareResponse;
-import org.springframework.stereotype.Service;
+import com.sesac.solbid.dto.payment.request.PaymentPrepareRequest;
+import com.sesac.solbid.dto.payment.response.PaymentPrepareResponse;
 
 
 public interface PaymentService {
