+++ backend/demo/src/main/java/com/sesac/solbid/service/PaymentServiceImpl.java
@@ -3,9 +3,9 @@ package com.sesac.solbid.service;
 import com.sesac.solbid.domain.Payments;
 import com.sesac.solbid.domain.User;
 import com.sesac.solbid.domain.enums.PaymentStatus;
-import com.sesac.solbid.dto.request.PaymentPrepareRequest;
-import com.sesac.solbid.dto.response.PaymentPrepareResponse;
-import com.sesac.solbid.dto.response.PortOnePaymentResponse;
+import com.sesac.solbid.dto.payment.request.PaymentPrepareRequest;
+import com.sesac.solbid.dto.payment.response.PaymentPrepareResponse;
+import com.sesac.solbid.dto.payment.response.PortOnePaymentResponse;
 import com.sesac.solbid.repository.PaymentsRepository;
 import com.sesac.solbid.repository.UserRepository;
 import jakarta.persistence.EntityManager;
