+++ backend/demo/src/main/java/com/sesac/solbid/service/PointService.java
@@ -1,6 +1,6 @@
 package com.sesac.solbid.service;
 
-import com.sesac.solbid.dto.response.PointSummaryResponse;
+import com.sesac.solbid.dto.payment.response.PointSummaryResponse;
 
 public interface PointService {
     PointSummaryResponse getCurrentPoint(Long userId);
